Save a preview image and load a mask for detection
An image of the view can be save in full resolution and 640 by 480.
The smaller image can be used to create a mask for areas where motion
detection is not needed. These areas are marked by painting them (using
paint or gimp etc.) black (color code 0). All other areas are set to 1
in software and the mask is mutiplied to the motion vetor data.